requirements.txt
customtkinter~=5.2.2
numpy~=1.19.5
pillow~=10.3.0
tensorflow~=2.5.0